Fix sensitive data exposure 2
